Move imported data stats code to own module
